Add support to pass in the entire x or y-axis into a Graph.

+++ java/src/test/java/gov/doc/ntia/sigmf/ext/global/sensor/NasctnSensorTest.java
@@ -90,9 +90,9 @@ public class NasctnSensorTest {
     Assertions.assertEquals("mean", series.get(1));
     Assertions.assertEquals(625L, psd.getLength().longValue());
     Assertions.assertEquals("dBm/Hz", psd.getYLabel());
-    Assertions.assertEquals(-5000000, psd.getXStart().get(0).intValue());
-    Assertions.assertEquals(16000, psd.getXStep().get(0).intValue());
-    Assertions.assertEquals(5000000, psd.getXStop().get(0).intValue());
+    Assertions.assertEquals(-5000000, ((Double)psd.getXStart().get(0)).intValue());
+    Assertions.assertEquals(16000, ((Double)psd.getXStep().get(0)).intValue());
+    Assertions.assertEquals(5000000, ((Double)psd.getXStop().get(0)).intValue());
     Assertions.assertEquals("fft", psd.getProcessing().get(0));
   }
